refactor: implement XDR checksum, React Query migration, Monaco editor leak cleanup and API rate limiting
